Fixes for bugs in code to let us run tests against LiteListener
build.gradle - Cleaning up declaration of listener dependency so it works
in Couchbase's environment

LiteTestCase - Get rid of logic that deleted test files from here to
context.

LiteTestContext - Allows caller to decide if subdirectory with data should
be deleted before test is run. By default the answer is yes.

Test9_LoadDB - The code assumed that DEFAULT_TEST_DB exists but since we
are re-using the same instance of LiteListener (due to TJWS bugs) we can't
rely on that so we create the DB if it's not there.